New design.
There's now an "overview" page, and a specific page for each
test. It looks nicer too :)

There are some internal changes for the sake of keeping track of
time and errors easily, including keeping a record of the average
test time, the 10 most recent errors (instead of just one), and
the overall pass/warn/error count per test.